refactor: widen x64-truncating pointer and handle fields (first-party C4311/C4312 to zero)
Stage 5 of the x64 port addresses every first-party compiler warning that
converts between a pointer and a 32-bit integer type. C4311 (pointer
truncation) drops from 24 sites to 0 and C4312 (narrow integer widened to
a pointer) drops from 27,697 sites to 1, taking the whole-tree warning
count from 27,809 to 19.

The bulk of the reduction comes from public/tier1/utlmemory.h, where
UTL_INVAL_SYMBOL was written as an int-sized sentinel that every pointer
debug cast in the codebase passed through; the literal is now (intp)
readable by all 21 memhandle call sites at once instead of each cast
re-narrowing a pointer at its own use site.

The remaining C4311 sites in the VGUI input ABI were genuine 64-bit
round-trip bugs rather than cosmetic truncation: IInput enumerated IME
language, conversion and sentence modes through int handleValue fields
while the underlying HIMC/HKL values are pointer-sized, so any consumer
read back a sign-extended low 32 bits on x64. Those fields, the two
getter returns and the three OnChangeIMEByHandle parameters are now
uintp throughout, and the matching TEXTENTRY_IME_* message handlers were
switched to the existing MESSAGE_FUNC_UINT64 / SetUint64 dispatch path
so KeyValues carries the full value end to end. uintp is already visible
to every one of those headers through vgui/VGUI.h, so no include had to
be added to the public ABI headers.

vguimatsurface/Input.cpp deliberately does not widen InputEvent_t.m_nData,
which remains an int as its name and every other reader assume it is;
instead the producer that was stuffing an HWND into that generic payload
now records it in a file-scope variable that the IE_IMESetWindow consumer
reads directly, which removes the round-trip without changing the
struct layout that other message types depend on. That variable and the
window procedure that writes it both sit inside the file's WIN32 block,
so the consumer forward is compiled under the same guard: the message is
posted nowhere else, which is why the Linux and macOS builds were failing
on an undeclared identifier before this guard was added.

SetMoveDone in baseentity.h only needed a width change, since converting
the int stored in m_hMoveDone to intp and from there to void* sign-extends
identically at all 43 call sites, so the fix is one cast rather than a new
field type. datamanager.h and gameui/Sys_Utils.h each had a single width
problem behind a whole class of call sites: INVALID_MEMHANDLE now uses
(uintp) so all 21 sites narrow consistently, and WHANDLE became intp so the
window procedure that takes it as its first argument matches the HWND
prototype ::EnumWindows actually passes.

engine/audio/voice_mixer_controls.cpp keeps a two-step (UINT)(UINT_PTR)
cast with a comment explaining it, because the Windows SDK
mixerGetDevCaps prototype genuinely takes UINT for a device id and the
value passing through it is never a pointer, so the two-step is the
honest description of that API rather than a laundering of a real
truncation. materialsystem/cmaterial.cpp keeps its (ITexture*)0xffffffff
sentinel untouched because the extension's reload semantics around that
magic value have not been audited, and it is one warning outside this
stage's gate.

lint-legacy's suspicious_ptr_cast ratchet moves from 45 to 29, with the
remaining 29 being sites where a pointer is intentionally narrowed for
comparison against a documented magic constant rather than a truncation
that could reach storage.

+++ game/server/baseentity.h
@@ -2654,7 +2654,7 @@ inline void CBaseEntity::FireBullets( int cShots, const Vector &vecSrc,
 	do \
 	{ \
 		m_pfnMoveDone = static_cast <void (CBaseEntity::*)(void)> (a); \
-		FunctionCheck( (void *)*((int *)((char *)this + ( offsetof(CBaseEntity,m_pfnMoveDone)))), "BaseMoveFunc" ); \
+		FunctionCheck( (void *)(intp) (*((int *)((char *)this + ( offsetof(CBaseEntity,m_pfnMoveDone))))), "BaseMoveFunc" ); \
 	} while ( 0 )
 #else
 #define SetMoveDone( a ) \
