+++ public/tier1/utlmemory.h
@@ -437,7 +437,7 @@ CUtlMemory<T,I>::~CUtlMemory()
 	Purge();
 
 #ifdef _DEBUG
-	m_pMemory = reinterpret_cast< T* >( 0xFEFEBAAD );
+	m_pMemory = reinterpret_cast< T* >( (intp)0xFEFEBAAD );
 	m_nAllocationCount = 0x7BADF00D;
 #endif
 }
